Addition of msg.payload.addtoheatmap and msg.payload.int/intensity
Addition of msg.payload.addtoheatmap and msg.payload.int/intensity.

To summarize.

added msg.payload.addtoheatmap. This defaults to true so you need to explicitly set it to false if you dont want that point to contribute to the heatmap. Done to make this change non breaking for other users.
added msg.payload.int or msg.payload.intensity (read notes below for when to use which one). This sets the intensity of the point you contribute to the heatmap. If you leave the heatmap max parameter at the default of 1 it will work with values of 0.1 - 1. if this parameter doesn't exist the current behavior as it stands today is applied.
Notes

there is a bug in leaflet-heat at the moment that if you change max parameter the intensity value doesn't work correctly (screws up when you zoom in and out). suggest leaving at default of 1 and scaling int/intensity between 0.1 - 1.
Looking through your code you have a few different formats that a positi